+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.3 March 19, 2024
 
 - Bug #11: Fix broken links in `docs` (@terabytesoftw)
+- Enh #12: Add `CheckboxList` and `RadioList` class widgets (@terabytesoftw)
 
 ## 0.1.2 March 15, 2024